move witnessgen
