+++ utils/client/src/client.rs
@@ -15,9 +15,6 @@ use kona_driver::DriverResult;
 use kona_driver::Executor;
 use kona_driver::TipCursor;
 use kona_executor::{KonaHandleRegister, TrieDBProvider};
-use kona_preimage::HintWriter;
-use kona_preimage::NativeChannel;
-use kona_preimage::OracleReader;
 use kona_preimage::{CommsClient, PreimageKey};
 use kona_proof::errors::OracleProviderError;
 use kona_proof::executor::KonaExecutor;
@@ -39,22 +36,6 @@ use tracing::info;
 use tracing::warn;
 
 use crate::oracle::OPSuccinctOracleBlobProvider;
-use crate::precompiles::zkvm_handle_register;
-use crate::InMemoryOracle;
-use crate::StoreOracle;
-
-pub async fn run_witnessgen_client(
-    preimage_chan: NativeChannel,
-    hint_chan: NativeChannel,
-) -> Result<InMemoryOracle> {
-    let oracle = Arc::new(StoreOracle::new(
-        OracleReader::new(preimage_chan),
-        HintWriter::new(hint_chan),
-    ));
-    let _ = run_opsuccinct_client(oracle.clone(), Some(zkvm_handle_register)).await?;
-    let in_memory_oracle = InMemoryOracle::populate_from_store(oracle.as_ref())?;
-    Ok(in_memory_oracle)
-}
 
 // Sourced from https://github.com/op-rs/kona/tree/main/bin/client/src/single.rs
 pub async fn run_opsuccinct_client<O>(
